test: move regressions into existing test files
